refactor(navigation): move files and shared styles into parent directory
