+++ src/app/infrastructure/shared/components/navigation/navigation.scss
@@ -10,7 +10,6 @@ button {
   flex-direction: column;
 }
 
-.profile-dropdown-container {
-  display: flex;
-  align-items: center;
+.navbar-nav {
+  align-self: baseline;
 }
