+++ src/app/infrastructure/shared/components/navigation/settings/settings.component.html
@@ -8,7 +8,7 @@
     Hi Username
   </button>
   <div ngbDropdownMenu aria-labelledby="profileDropdown">
-    <div ngbDropdownItem class="profile-dropdown-container">
+    <div class="profile-dropdown-container">
       <span>Name</span>
       <div>
         <ng-container *ngFor="let link of profileLinks; trackBy: trackByLink">
