+++ src/app/infrastructure/shared/components/navigation/settings/settings.component.scss
@@ -1,0 +1,6 @@
+@import '../navigation.scss';
+
+.profile-dropdown-container {
+  display: flex;
+  align-items: center;
+}
