+++ src/app/infrastructure/shared/components/navigation/side-navigation/side-navigation.component.ts
