+++ src/app/infrastructure/shared/components/navigation/top-navigation/top-navigation.component.scss
@@ -1,0 +1,1 @@
+@import '../navigation.scss';
