--- src/app/infrastructure/shared/components/top-navigation/top-navigation.component.scss
@@ -1,20 +1,0 @@
-button {
-  cursor: pointer;
-  border: none;
-  background-color: transparent;
-}
-
-.menu-container {
-  padding-right: 20px;
-  height: 100%;
-}
-
-.nav-link {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.navbar-nav {
-  align-self: baseline;
-}
